* master:
  Add a .travis.yml (#7)

+++ test/integration/memcacheserver.js
@@ -3,9 +3,14 @@ var child_process = require('child_process')
 var available_port = 11234
 
 function MemcacheServer(port) {
-  this.port = port || available_port++
   this.process = null
-  this.listen()
+  if (process.env.MEMCACHING_TEST_HOST) {
+    this.host = process.env.MEMCACHING_TEST_HOST
+  } else {
+    this.port = port || available_port++
+    this.host = '127.0.0.1:' + this.port
+    this.listen()
+  }
 }
 
 MemcacheServer.prototype = {}
@@ -20,7 +25,9 @@ MemcacheServer.prototype.listen = function () {
 }
 
 MemcacheServer.prototype.end = function () {
-  this.process.kill()
+  if (this.process) {
+    this.process.kill()
+  }
   this.process = null
 }
